don't get confused by cabal.project files
Fixes #52.

When run from a folder "<my-project>" which contains a cabal.project
file, policeman will fail with the following message:

  cabal: No targets given and there is no package in the current
  directory. [...]

In fact, there is a package in the current directory, which policeman
just created at "<my-project>/.policeman-evidence/<my-project>-<previous-version>".
Unfortunately, since that newly-created directory is not listed in
"<my-project>/cabal.project", it is ignored by cabal.

The solution is to create a competing file at
"<my-project>/.policeman-evidence/<my-project>-<previous-version>/cabal.project".
cabal will find that file first, and since that file does list the
current directory, cabal will accept to build it.